requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
idna==3.10
psycopg2==2.9.9
PyJWT==2.9.0
redis==3.5.3
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
urllib3==2.2.3